refactored and works
