+++ src/components/common/Logout/index.ts
@@ -1,0 +1,1 @@
+export { default as Logout } from './Logout';
