--- src/components/common/LogoutPopup/index.ts
@@ -1,1 +1,0 @@
-export { default as LogoutPopup } from './LogoutPopup';
