+++ src/styles/components/common/logout.scss
@@ -1,0 +1,9 @@
+.logout {
+  p {
+    font-family: $text-font;
+    color: $white;
+    font-size: 1.2rem;
+    // margin-bottom: 2rem;
+    cursor: pointer;
+  }
+}
