--- src/styles/components/common/logoutPopup.scss
@@ -1,21 +1,0 @@
-.logout-popup {
-  p {
-    font-family: $text-font;
-    color: $white;
-    font-size: 1.6rem;
-    margin-bottom: 2rem;
-  }
-  .button-row {
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: space-evenly;
-    button.no-button {
-      @include button($red);
-      padding: 0.7rem 1rem;
-    }
-    button {
-      width: 40%;
-      padding: 0.7rem 1rem;
-    }
-  }
-}
