+++ src/styles/index.scss
@@ -17,7 +17,7 @@
 // Common Component Stylesheets
 @import './components/common/input.scss';
 @import './components/common/modal.scss';
-@import './components/common/logoutPopup.scss';
+@import './components/common/logout.scss';
 @import './components/common/checkbox.scss';
 @import './components/common/landingPageWrapper.scss';
 @import './components/common/cleverButton.scss';
